requirements/base.txt
Django>=1.8
django-environ>=0.3.0
django-braces>=1.8.0
django-crispy-forms>=1.4.0
django-admin-bootstrapped>=2.4.0
django-autocomplete-light==2.2.5
django-timezone-field>=1.2 
django-reversion==1.9.1
django-authtools>=1.2.0
django-compressor>=1.5
django-appconf>=1.0
easy-thumbnails>=2.2
Pillow>=2.9.0
argparse>=1.2.1
gunicorn>=19.3.0
requests>=2.7.0
requests-futures>=0.9.5
pytz>=2015.4
six>=1.9.0
sqlparse>=0.1.16
